@@ -1,9 +1,13 @@
 module github.com/tutorialedge/go-grpc-tutorial
 
-go 1.13
+go 1.14
 
 require (
-	github.com/golang/protobuf v1.3.3
-	golang.org/x/net v0.0.0-20190311183353-d8887717615a
+	github.com/golang/protobuf v1.4.1
+	golang.org/x/net v0.0.0-20200506145744-7e3656a0809f
+	golang.org/x/sys v0.0.0-20200509044756-6aff5f38e54f // indirect
+	golang.org/x/text v0.3.2 // indirect
+	google.golang.org/genproto v0.0.0-20200511104702-f5ebc3bea380 // indirect
 	google.golang.org/grpc v1.29.1
+	google.golang.org/protobuf v1.22.0
 )
